backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
SQLAlchemy==2.0.31
asyncpg==0.29.0
alembic==1.13.2
python-jose[cryptography]==3.3.0
httpx==0.27.0
pydantic==2.7.1
pydantic-settings==2.3.0
python-multipart==0.0.9
